tests/requirements.txt
pyzbar
pillow